Add increment: Level-6
Implemented `delete` function
